Implemented Trade Screen Persistence
When the trade screen was refreshed (such as after a trade offer), the selected counterparty and horse were reset back to the first value in the list. This has been changed so that there is memory of the previous selection. Also, when switching between counterparties, the most recent horse selected for a counterparty is remembered and automatically highlighted if that counterparty is selected in the future.