Changed parameter loading to check for local versions
Previously, any changes made to a parameter file would result in those changes being included in the git commit. Now you can append "local_" to the name of a parameter file and that file will be preferentially loaded. Just don't add it to the commit or that defeats the purpose. For instance local_genes.json will be loaded. If that file doesn't exist, genes.json will be loaded instead.

Also consolidated and refactored the various parameter files. As part of this, the horse_names file is now a json with catagories for male, female, and unisex names.