Added injuries.
Horses can injure themselves in races. The types of injuries are defined in injuries.json. Injuries effect different body parts and cause different amounts of damage. Damage heals over time. If a horse is too injured, it will not be able to race. Some injuries are fatal.

Made the message display scrolling. It will now keep a certain number of old messages.

Made a debug window that shows a horse's properties and injuries.