Better horse history generation
From the main screen, the player can now select different options for starting the game. Either 0, 10, or 20 years of horse history simulation. During the simulation, a starting population of 50 horses is bred and raced. Injuries don't occur but death by old age does.

There is also a db file from a 20 year simulation that is saved and can be loaded from the same menu dropdown.