Introduced real saving/loading functionality.
The player now has the ability to save and load games. To accomplish this, all game data is stored in the sqlite database. This was already being done for most aspects, but now estate information is completely in the database (rather than in a transient object). Date information is now stored in the database. The database is no longer kept in memory because there was no significant speed advantage to this.

To save the game, the active database file is copied to the desired filename. To load the game, the desired file is copied as 'active_game.db' and then connected to.