Added a calendar to the game
Annual events (just races for now) can be defined in events.json. The parameters for this race (name, distance, purse, etc.) are all set in this json. These events get added to the calendar table on December 31st for the whole next year. Every day, the calendar table is checked to see if there are any events happening and these are executed. Right now, only races are executed but other events could be added in the future.

The race execution code has also been improved. When setting up a race, a random name can be generated. Additionally, a new parameter called speed_bonus has been added. This gives the AI horses a fixed speed boost which creates a harder race for the player.