Added new ways to see horse properties.
There are now ways to view a horse's weight, speed, and heart size. The player must purchase equipment (in the building tab) to unlock these. The idea is to add more of these as appropriate.

As part of this, there are some new building improvements. A tooltip text can be specified as 'text'. A maximum number purchasable can be specified as 'max'.